+++ src/it/scala/ai/tripl/arc/transform/PostgresDebeziumTransformSuite.scala
@@ -22,6 +22,7 @@ import ai.tripl.arc.api._
 import ai.tripl.arc.api.API._
 import ai.tripl.arc.util.log.LoggerFactory
 import ai.tripl.arc.udf.UDF
+import ai.tripl.arc.transform.DebeziumStringKafkaEvent
 
 import ai.tripl.arc.util._
 
@@ -31,6 +32,7 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
 
   val inputView = "inputView"
   val outputView = "outputView"
+  val initialStateView = "initialStateView"
   val schema = "schema"
   val checkpointLocation = "/tmp/debezium"
 
@@ -45,6 +47,8 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
                   .config("spark.ui.port", "4040")
                   .config("spark.checkpoint.compress", "true")
                   .config("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
+                  .config("spark.kryo.unsafe", "true")
+                  .config("spark.kryoserializer.buffer.max", "1024m")
                   .config("spark.sql.streaming.checkpointLocation", checkpointLocation)
                   .config("spark.sql.streaming.forceDeleteTempCheckpointLocation", "true")
                   .appName("Arc Test")
@@ -67,7 +71,7 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
         println("Query terminated: " + queryTerminated.id)
       }
       override def onQueryProgress(queryProgress: QueryProgressEvent): Unit = {
-        println(s"numRowsTotal: ${queryProgress.progress.stateOperators(0).numRowsTotal} inputRowsPerSecond: ${queryProgress.progress.inputRowsPerSecond.round} processedRowsPerSecond: ${queryProgress.progress.processedRowsPerSecond.round}")
+        println(s"numRowsTotal: ${if (queryProgress.progress.stateOperators.length == 0) 0 else queryProgress.progress.stateOperators(0).numRowsTotal} inputRowsPerSecond: ${queryProgress.progress.inputRowsPerSecond.round} processedRowsPerSecond: ${queryProgress.progress.processedRowsPerSecond.round}")
       }
     })
 
@@ -93,15 +97,16 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
     |    "database.server.name": "dbserver1",
     |    "database.history.kafka.bootstrap.servers": "kafka:9092",
     |    "database.history.kafka.topic": "schema-changes.inventory",
-    |    "schema.whitelist": "inventory",
+    |    "schema.include.list": "inventory",
     |    "message.key.columns": "${tableName}:${key}",
-    |    "decimal.handling.mode": "string"
+    |    "decimal.handling.mode": "string",
+    |    "heartbeat.interval​.ms": 20
     |  }
     |}""".stripMargin
   }
 
   def makeTransaction(statements: Seq[String]): String = {
-    s"""START TRANSACTION;
+    s"""BEGIN TRANSACTION ISOLATION LEVEL SERIALIZABLE;
     |${statements.mkString("\n")}
     |COMMIT;
     """.stripMargin
@@ -185,7 +190,7 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
     (transactions, updates, inserts, deletes)
   }
 
-  test("PostgresDebeziumTransform") {
+  test("PostgresDebeziumTransform: Streaming") {
     implicit val spark = session
     import spark.implicits._
     implicit val logger = TestUtils.getLogger()
@@ -210,7 +215,7 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
             description=None,
             inputURI=new URI(postgresURL),
             jdbcURL=postgresURL,
-            sql=makeTransaction(Seq(s"CREATE TABLE ${tableName} (c_custkey INTEGER PRIMARY KEY NOT NULL, c_name VARCHAR(25) NOT NULL, c_address VARCHAR(40) NOT NULL, c_nationkey INTEGER NOT NULL, c_phone VARCHAR(15) NOT NULL, c_acctbal DECIMAL(20,2) NOT NULL, c_mktsegment VARCHAR(10) NOT NULL, c_comment VARCHAR(117) NOT NULL);")),
+            sql=makeTransaction(Seq(s"CREATE TABLE inventory.${tableName} (c_custkey INTEGER PRIMARY KEY NOT NULL, c_name VARCHAR(25) NOT NULL, c_address VARCHAR(40) NOT NULL, c_nationkey INTEGER NOT NULL, c_phone VARCHAR(15) NOT NULL, c_acctbal DECIMAL(20,2) NOT NULL, c_mktsegment VARCHAR(10) NOT NULL, c_comment VARCHAR(117) NOT NULL);")),
             params=Map.empty,
             sqlParams=Map.empty
           )
@@ -240,7 +245,12 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
             inputView=inputView,
             outputView=outputView,
             schema=Left(schema),
-            strict=strict
+            strict=strict,
+            initialStateView=None,
+            initialStateKey=None,
+            persist=true,
+            numPartitions=None,
+            partitionBy=List.empty,
           )
         )
 
@@ -285,6 +295,7 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
               )
             )
           }
+          println(s"executed ${transactions.length} transactions against ${tableName} with ${update} updates, ${insert} inserts, ${delete} deletes")
 
           writeStream.processAllAvailable
           writeStream.stop
@@ -312,8 +323,8 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
             )
           ).get
           assert(TestUtils.datasetEquality(expected, spark.table(tableName)))
+          println("PASS\n")
 
-          println(s"executed ${transactions.length} transactions against ${tableName} with ${update} updates, ${insert} inserts, ${delete} deletes\n")
         } catch {
           case e: Exception => fail(e.getMessage)
         } finally {
@@ -359,7 +370,7 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
         description=None,
         inputURI=new URI(postgresURL),
         jdbcURL=postgresURL,
-        sql=makeTransaction(Seq(s"CREATE TABLE ${tableName} (booleanDatum BOOLEAN NOT NULL, dateDatum DATE NOT NULL, decimalDatum DECIMAL(10,3) NOT NULL, doubleDatum DOUBLE PRECISION NOT NULL, integerDatum INTEGER NOT NULL, longDatum BIGINT NOT NULL, stringDatum VARCHAR(255) NOT NULL, timeDatum VARCHAR(255) NOT NULL, timestampDatum TIMESTAMP NOT NULL);")),
+        sql=makeTransaction(Seq(s"CREATE TABLE inventory.${tableName} (booleanDatum BOOLEAN NOT NULL, dateDatum DATE NOT NULL, decimalDatum DECIMAL(10,3) NOT NULL, doubleDatum DOUBLE PRECISION NOT NULL, integerDatum INTEGER NOT NULL, longDatum BIGINT NOT NULL, stringDatum VARCHAR(255) NOT NULL, timeDatum VARCHAR(255) NOT NULL, timestampDatum TIMESTAMP NOT NULL);")),
         params=Map.empty,
         sqlParams=Map.empty
       )
@@ -387,7 +398,12 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
         inputView=inputView,
         outputView=outputView,
         schema=Left(schema),
-        strict=true
+        strict=true,
+        initialStateView=None,
+        initialStateKey=None,
+        persist=true,
+        numPartitions=None,
+        partitionBy=List.empty,
       )
     )
 
@@ -430,6 +446,188 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
         )
       )
       writeStream.stop
+    }
+  }
+
+  test("PostgresDebeziumTransform: Batch") {
+    implicit val spark = session
+    import spark.implicits._
+    implicit val logger = TestUtils.getLogger()
+    implicit val arcContext = TestUtils.getARCContext(isStreaming = true)
+    val size = 10
+
+    val (customerInitial, customerUpdates) = TestHelpers.getTestData(size)
+    val customersMetadata = MetadataUtils.createMetadataDataframe(customerInitial.toDF)
+    customersMetadata.persist
+    customersMetadata.createOrReplaceTempView(schema)
+
+    println()
+    for (seed <- 0 to 0) {
+      for (strict <- Seq(true)) {
+        val tableName = s"customers_${UUID.randomUUID.toString.replaceAll("-","")}"
+        println(s"postgres ${if (strict) "strict" else "not-strict"} seed: ${seed} target: ${tableName}")
+
+        ai.tripl.arc.execute.JDBCExecuteStage.execute(
+          ai.tripl.arc.execute.JDBCExecuteStage(
+            plugin=new ai.tripl.arc.execute.JDBCExecute,
+            id=None,
+            name="JDBCExecute",
+            description=None,
+            inputURI=new URI(postgresURL),
+            jdbcURL=postgresURL,
+            sql=makeTransaction(Seq(s"CREATE TABLE inventory.${tableName} (c_custkey INTEGER PRIMARY KEY NOT NULL, c_name VARCHAR(25) NOT NULL, c_address VARCHAR(40) NOT NULL, c_nationkey INTEGER NOT NULL, c_phone VARCHAR(15) NOT NULL, c_acctbal DECIMAL(20,2) NOT NULL, c_mktsegment VARCHAR(10) NOT NULL, c_comment VARCHAR(117) NOT NULL);")),
+            params=Map.empty,
+            sqlParams=Map.empty
+          )
+        )
+        customerInitial.write.mode("append").jdbc(postgresURL, s"inventory.${tableName}", new java.util.Properties)
+
+        // make transactions
+        val (transactions, update, insert, delete) = makeTransactions(customerInitial, customerUpdates, s"inventory.${tableName}", seed)
+
+        // while running perform PARALLEL insert/update/delete transactions
+        // this will block the main thread but we want to process all updates before triggering awaitTermination
+        var last = System.currentTimeMillis()
+        var i = 0
+        transactions.par.foreach { sql =>
+          println(sql)
+          if (System.currentTimeMillis() > last+1000) {
+            last = System.currentTimeMillis()
+            println(s"${i} transactions/sec")
+            i = 0
+          }
+          i += 1
+          ai.tripl.arc.execute.JDBCExecuteStage.execute(
+            ai.tripl.arc.execute.JDBCExecuteStage(
+              plugin=new ai.tripl.arc.execute.JDBCExecute,
+              id=None,
+              name="JDBCExecute",
+              description=None,
+              inputURI=new URI(postgresURL),
+              jdbcURL=postgresURL,
+              sql=sql,
+              params=Map.empty,
+              sqlParams=Map.empty
+            )
+          )
+        }
+        println(s"executed ${transactions.length} transactions against ${tableName} with ${update} updates, ${insert} inserts, ${delete} deletes")
+
+        TestHelpers.registerConnector(connectURI, makeConnectorConfig(s"inventory.${tableName}", "c_custkey"))
+
+        val readStream = spark
+          .readStream
+          .format("kafka")
+          .option("kafka.bootstrap.servers", "kafka:9092")
+          .option("subscribe", s"dbserver1.inventory.${tableName}")
+          .option("startingOffsets", "earliest")
+          .load
+
+        val writeStream = readStream
+          .writeStream
+          .outputMode("append")
+          .queryName(tableName)
+          .format("memory")
+          .start
+
+        try {
+          // wait for query to start
+          val start = System.currentTimeMillis()
+          while (writeStream.lastProgress == null || (writeStream.lastProgress != null && writeStream.lastProgress.numInputRows == 0)) {
+            if (System.currentTimeMillis() > start + 180000) throw new Exception("Timeout without messages arriving")
+            println("Waiting for query progress...")
+            Thread.sleep(1000)
+          }
+
+          writeStream.processAllAvailable
+          writeStream.stop
+
+          // read in batch mode
+          val read = spark
+            .read
+            .format("kafka")
+            .option("kafka.bootstrap.servers", "kafka:9092")
+            .option("subscribe", s"dbserver1.inventory.${tableName}")
+            .option("startingOffsets", "earliest")
+            .load
+            .as[DebeziumStringKafkaEvent]
+            .collect
+            .sortBy(row => row.offset)
+            .toList
+
+          // recursively apply the records passing in the previous state
+          println(s"processing ${read.size} events...")
+          read.grouped(size/3).zipWithIndex.foreach { case (batch, index) =>
+            batch.toDF.createOrReplaceTempView(inputView)
+            transform.DebeziumTransformStage.execute(
+              transform.DebeziumTransformStage(
+                plugin=new transform.DebeziumTransform,
+                id=None,
+                name="DebeziumTransform",
+                description=None,
+                inputView=inputView,
+                outputView=outputView,
+                schema=Left(schema),
+                strict=strict,
+                initialStateView=if (index == 0) None else Option(initialStateView),
+                initialStateKey=Option("c_custkey"),
+                persist=false,
+                numPartitions=None,
+                partitionBy=List.empty,
+              )
+            )
+
+            val output = spark.table(outputView)
+            output.createOrReplaceTempView(initialStateView)
+          }
+
+          // validate results
+          val expected = extract.JDBCExtractStage.execute(
+            extract.JDBCExtractStage(
+              plugin=new extract.JDBCExtract,
+              id=None,
+              name="dataset",
+              description=None,
+              schema=Right(Nil),
+              outputView="expected",
+              jdbcURL=postgresURL,
+              driver=DriverManager.getDriver(postgresURL),
+              tableName=s"inventory.${tableName}",
+              numPartitions=None,
+              fetchsize=None,
+              partitionBy=Nil,
+              customSchema=None,
+              persist=true,
+              partitionColumn=None,
+              predicates=Nil,
+              params=Map.empty
+            )
+          ).get
+
+          assert(TestUtils.datasetEquality(expected, spark.table(outputView)))
+          println(s"PASS: expected: ${expected.count} actual: ${spark.table(outputView).count}\n")
+
+        } catch {
+          case e: Exception => fail(e.getMessage)
+        } finally {
+          TestHelpers.deleteConnector(connectURI, connectorName)
+          TestHelpers.deleteConnector(connectURI, s"${connectorName}-dbhistory")
+          // ai.tripl.arc.execute.JDBCExecuteStage.execute(
+          //   ai.tripl.arc.execute.JDBCExecuteStage(
+          //     plugin=new ai.tripl.arc.execute.JDBCExecute,
+          //     id=None,
+          //     name="JDBCExecute",
+          //     description=None,
+          //     inputURI=new URI(postgresURL),
+          //     jdbcURL=postgresURL,
+          //     sql=makeTransaction(Seq(s"DROP TABLE inventory.${tableName};")),
+          //     params=Map.empty,
+          //     sqlParams=Map.empty
+          //   )
+          // )
+          writeStream.stop
+        }
+      }
     }
   }