+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.1.0
+
+- remove constraint to only support `streaming` mode.
+- add support for injecting a previous state by providing `initialStateView` and `initialStateKey`.
+
 ## 1.0.0
 
 - initial release.